+++ docs/source/conf.py
@@ -25,9 +25,9 @@ copyright = '2019, Geoffrey M. Sametz'
 author = 'Geoffrey M. Sametz'
 
 # The short X.Y version
-version = '0.2'
+version = '0.3'
 # The full version, including alpha/beta/rc tags
-release = '0.2.2-alpha'
+release = '0.3.0-beta'
 
 
 # -- General configuration ---------------------------------------------------