+++ nmrsim/__init__.py
@@ -62,4 +62,4 @@ implementation).
 
 from ._classes import Multiplet, SpinSystem, Spectrum  # noqa: F401
 
-__version__ = '0.2.2'
+__version__ = '0.3.0'