+++ CHANGELOG.rst
@@ -23,6 +23,32 @@ Working towards a Version 1.0.0 release, the author interprets the terms below a
 * **Version 1.0.0 release**: API is stable. The package is available on PyPI (and perhaps conda).
 
 
+Unreleased
+----------
+Added
+^^^^^
+* nmrsim.Multiplet:
+    * :code:`w` attribute added (peak width at half height).
+* nmrsim.SpinSystem:
+    * :code:`w` attribute added (peak width at half height).
+* nmrsim.Spectrum:
+    * In-place addition modifies the Spectrum object in-place.
+    * :code:`vmin` and :code:`vmax` attributes added, to set spectral width.
+    * :code:`default_limits` method added, to reset spectral width to default.
+    * :code:`lineshape()` method added, to return lineshape data for the
+      spectrum.
+
+Changed
+^^^^^^^
+* nmrsim.firstorder.multiplet now returns a *sorted* peaklist.
+
+Fixed
+^^^^^
+* nmrsim.Spectum:
+    * Addition of Spectrum objects is commutative, and returns a new Spectrum
+      object.
+
+
 0.2.1 - 2019-07-26 (alpha release)
 ----------------------------------
 * Binaries for accelerating qm calculations packaged in distribution.